requirements.txt
torch==2.1.0
transformers==4.32.0
bitsandbytes==0.40.0
datasets==2.14.0
peft==0.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.98.0
uvicorn==0.23.0
shap==0.41.0
pydantic==1.10.12
